+++ packages/twenty-front/src/modules/workflow/search-variables/types/WorkflowStepMock.ts
@@ -1,0 +1,5 @@
+export type WorkflowStepMock = {
+  id: string;
+  name: string;
+  output: Record<string, any>;
+};
